requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.1.2
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==23.1.0
black==23.11.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
dataclasses==0.6
docopt==0.6.2
fastjsonschema==2.19.0
filelock==3.13.1
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.45.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
hydra-core==1.3.2
idna==3.6
imageio==2.33.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==6.8.0
imutils==0.5.4
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_core==5.5.1
keras==2.15.0
kiwisolver==1.4.5
kmedoids==0.5.0
labelme2coco==0.2.4
lazy_loader==0.3
libclang==16.0.6
lime==0.2.0.1
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbformat==5.9.2
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
pigar==2.1.1
Pillow==10.1.0
pipreqs==0.4.13
pixellib==0.7.1
platformdirs==4.1.0
portalocker==2.8.2
protobuf==4.23.4
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
sahi==0.11.15
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
streamlit==1.28.2
streamlit-cropper==0.2.2
streamlit-option-menu==0.3.6
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
xmljson==0.2.1
yacs==0.1.8
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
